исправил деление

+++ src/org/Ex2/Calculator.java
@@ -67,7 +67,7 @@ package org.Ex2;
         System.out.println(lResult);
     }
     public void split(long longNum1, long longNum2){
-        double longResult = longNum1/longNum2;
+        long longResult = longNum1/longNum2;
         System.out.println(longResult);
     }
